copy ROM into vm.m (main memory) rather than keep reference
ROMs are not strictly speaking Read-Only since at the very least the
contents of the PC will change during the course of execution and the
PC is just the first word of main memory. So in order to be able to
run a ROM more than one, copy it when constructing a MakoVM object.